+++ pages/authLogin/authLogin.js
@@ -21,10 +21,11 @@ router.get('/', function(req, res, next) {
     sqldb.query(sql.get_mode, params, function(err, result) {
         if (ERR(err, next)) return;
         res.locals.mode = result.rows[0].mode;
-        if (res.locals.mode == 'Exam') {
-            res.locals.config.hasOauth = false;
-            res.locals.config.hasAzure = false;
-        }
+        // We could set res.locals.config.hasOauth = false (or
+        // hasAzure) to not display those options inside the CBTF, but
+        // this will also need to depend on which institution we have
+        // detected (e.g., UIUC doesn't want Azure during exams, but
+        // ZJUI does want it).
         res.render(__filename.replace(/\.js$/, '.ejs'), res.locals);
     });
 });